// src/lib/api/axios/types.ts
export interface ClientError {
  errorCode: number;
  errorMsg: string;
}
// ...
export interface SiteDBInfo extends ReqPostSiteDBInfo {
  index: number; // Create an index after receiving the response.
  crasCompanyFabName: string;
}

export interface ReqPostSiteDBInfo {
  siteId: number;
  crasCompanyName: string;
  crasFabName: string;
  crasAddress: string;
  crasPort: number;
  emailAddress: string;
  emailPort: number;
  emailUserName: string;
  emailPassword: string;
  emailFrom: string;
  rssAddress: string;
  rssPort: number;
  rssUserName: string;
  rssPassword: string;
}

export interface ReqPutSiteDBInfo extends ReqPostSiteDBInfo {
  siteId: number;
}

export interface ReqPostCrasConnection {
  crasAddress: string;
  crasPort: number;
}

export interface ReqPostEmailConnection {
  emailAddress: string;
  emailPort: number;
  emailUserName: string;
  emailPassword: string;
}

export interface ReqPostRssConnection {
  rssAddress: string;
  rssPort: number;
  rssUserName: string;
  rssPassword: string;
  crasAddress: string;
  crasPort: number;
}